requirements.txt
numpy
pandas
python-Levenshtein
sklearn
matplotlib